+++ base_layer/core/tests/node_comms_interface.rs
@@ -38,7 +38,7 @@ use tari_core::{
     transactions::{
         tari_amount::MicroTari,
         test_helpers::{create_utxo, spend_utxos},
-        transaction::{OutputFeatures, TransactionOutput, UnblindedOutput},
+        transaction_entities::{OutputFeatures, TransactionOutput, UnblindedOutput},
         CryptoFactories,
     },
     txn_schema,