+++ base_layer/wallet/tests/transaction_service/service.rs
@@ -91,7 +91,7 @@ use tari_core::{
         fee::Fee,
         tari_amount::*,
         test_helpers::{create_unblinded_output, TestParams as TestParamsHelpers},
-        transaction::{KernelBuilder, KernelFeatures, OutputFeatures, Transaction},
+        transaction_entities::{KernelBuilder, KernelFeatures, OutputFeatures, Transaction},
         transaction_protocol::{proto, recipient::RecipientSignedMessage, sender::TransactionSenderMessage},
         CryptoFactories,
         ReceiverTransactionProtocol,