+++ comms/src/protocol/messaging/outbound.rs
@@ -20,7 +20,7 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use super::{error::MessagingProtocolError, MessagingEvent, MessagingProtocol, SendFailReason};
+use super::{error::MessagingProtocolError, metrics, MessagingEvent, MessagingProtocol, SendFailReason};
 use crate::{
     connection_manager::{NegotiatedSubstream, PeerConnection},
     connectivity::{ConnectivityError, ConnectivityRequester},
@@ -71,6 +71,7 @@ impl OutboundMessaging {
             "comms::messaging::outbound",
             node_id = self.peer_node_id.to_string().as_str()
         );
+        metrics::num_sessions().inc();
         async move {
             debug!(
                 target: LOG_TARGET,
@@ -112,6 +113,7 @@ impl OutboundMessaging {
                     );
                 },
                 Err(err) => {
+                    metrics::error_count(&peer_node_id).inc();
                     error!(
                         target: LOG_TARGET,
                         "Outbound messaging protocol failed for peer {}: {}", peer_node_id, err
@@ -119,6 +121,7 @@ impl OutboundMessaging {
                 },
             }
 
+            metrics::num_sessions().dec();
             let _ = messaging_events_tx
                 .send(MessagingEvent::OutboundProtocolExited(peer_node_id))
                 .await;
@@ -291,7 +294,9 @@ impl OutboundMessaging {
             None => Either::Right(stream.map(Ok)),
         };
 
+        let outbound_count = metrics::outbound_message_count(&self.peer_node_id);
         let stream = stream.map(|msg| {
+            outbound_count.inc();
             msg.map(|mut out_msg| {
                 event!(Level::DEBUG, "Message buffered for sending {}", out_msg);
                 out_msg.reply_success();