+++ comms/src/protocol/rpc/client/metrics.rs
@@ -22,69 +22,108 @@
 
 use crate::{peer_manager::NodeId, protocol::ProtocolId};
 use once_cell::sync::Lazy;
-use tari_metrics::{Histogram, HistogramVec, IntGauge, IntGaugeVec};
+use tari_metrics::{Histogram, HistogramVec, IntCounter, IntCounterVec, IntGauge, IntGaugeVec};
 
-pub fn sessions_counter(node_id: &NodeId, protocol: &ProtocolId) -> IntGauge {
+pub fn num_sessions(peer: &NodeId, protocol: &ProtocolId) -> IntGauge {
     static METER: Lazy<IntGaugeVec> = Lazy::new(|| {
         tari_metrics::register_int_gauge_vec(
             "comms::rpc::client::num_sessions",
-            "The number of active clients per node per protocol",
-            &["peer", "protocol"],
+            "The number of active clients per peer per protocol",
+            &["peer_id", "protocol"],
         )
         .unwrap()
     });
 
-    METER.with_label_values(&[node_id.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
+    METER.with_label_values(&[peer.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
 }
 
-pub fn handshake_errors(node_id: &NodeId, protocol: &ProtocolId) -> IntGauge {
-    static METER: Lazy<IntGaugeVec> = Lazy::new(|| {
-        tari_metrics::register_int_gauge_vec(
+pub fn handshake_counter(peer: &NodeId, protocol: &ProtocolId) -> IntCounter {
+    static METER: Lazy<IntCounterVec> = Lazy::new(|| {
+        tari_metrics::register_int_counter_vec(
+            "comms::rpc::client::handshake_count",
+            "The number of handshakes per peer per protocol",
+            &["peer_id", "protocol"],
+        )
+        .unwrap()
+    });
+
+    METER.with_label_values(&[peer.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
+}
+
+pub fn handshake_errors(peer: &NodeId, protocol: &ProtocolId) -> IntCounter {
+    static METER: Lazy<IntCounterVec> = Lazy::new(|| {
+        tari_metrics::register_int_counter_vec(
             "comms::rpc::client::handshake_errors",
-            "The number of handshake errors per node per protocol",
-            &["peer", "protocol"],
+            "The number of handshake errors per peer per protocol",
+            &["peer_id", "protocol"],
         )
         .unwrap()
     });
 
-    METER.with_label_values(&[node_id.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
+    METER.with_label_values(&[peer.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
 }
 
-pub fn request_response_latency(node_id: &NodeId, protocol: &ProtocolId) -> Histogram {
+pub fn client_errors(peer: &NodeId, protocol: &ProtocolId) -> IntCounter {
+    static METER: Lazy<IntCounterVec> = Lazy::new(|| {
+        tari_metrics::register_int_counter_vec(
+            "comms::rpc::client::error_count",
+            "The number of client errors per peer per protocol",
+            &["peer_id", "protocol"],
+        )
+        .unwrap()
+    });
+
+    METER.with_label_values(&[peer.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
+}
+
+pub fn client_timeouts(peer: &NodeId, protocol: &ProtocolId) -> IntCounter {
+    static METER: Lazy<IntCounterVec> = Lazy::new(|| {
+        tari_metrics::register_int_counter_vec(
+            "comms::rpc::client::error_timeouts",
+            "The number of client timeouts per peer per protocol",
+            &["peer_id", "protocol"],
+        )
+        .unwrap()
+    });
+
+    METER.with_label_values(&[peer.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
+}
+
+pub fn request_response_latency(peer: &NodeId, protocol: &ProtocolId) -> Histogram {
     static METER: Lazy<HistogramVec> = Lazy::new(|| {
         tari_metrics::register_histogram_vec(
             "comms::rpc::client::request_response_latency",
             "A histogram of request to first response latency",
-            &["peer", "protocol"],
+            &["peer_id", "protocol"],
         )
         .unwrap()
     });
 
-    METER.with_label_values(&[node_id.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
+    METER.with_label_values(&[peer.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
 }
 
-pub fn outbound_request_bytes(node_id: &NodeId, protocol: &ProtocolId) -> Histogram {
+pub fn outbound_request_bytes(peer: &NodeId, protocol: &ProtocolId) -> Histogram {
     static METER: Lazy<HistogramVec> = Lazy::new(|| {
         tari_metrics::register_histogram_vec(
             "comms::rpc::client::outbound_request_bytes",
-            "Avg. request bytes per node per protocol",
-            &["peer", "protocol"],
+            "Avg. request bytes per peer per protocol",
+            &["peer_id", "protocol"],
         )
         .unwrap()
     });
 
-    METER.with_label_values(&[node_id.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
+    METER.with_label_values(&[peer.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
 }
 
-pub fn inbound_response_bytes(node_id: &NodeId, protocol: &ProtocolId) -> Histogram {
+pub fn inbound_response_bytes(peer: &NodeId, protocol: &ProtocolId) -> Histogram {
     static METER: Lazy<HistogramVec> = Lazy::new(|| {
         tari_metrics::register_histogram_vec(
             "comms::rpc::client::inbound_response_bytes",
             "Avg. response bytes per peer per protocol",
-            &["peer", "protocol"],
+            &["peer_id", "protocol"],
         )
         .unwrap()
     });
 
-    METER.with_label_values(&[node_id.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
+    METER.with_label_values(&[peer.to_string().as_str(), String::from_utf8_lossy(protocol).as_ref()])
 }